fix launch and asset config

+++ primitives/core/src/lib.rs
@@ -59,7 +59,7 @@ mod types {
 	pub type Amount = i128;
 
 	/// AssetId type
-	pub type AssetId = u32;
+	pub type AssetId = u128;
 
 	/// Balance of an account.
 	pub type Balance = u128;
